add VS 2017

+++ source/lib/iconv.c
@@ -429,7 +429,7 @@ void iconvlist (int (*do_one) (unsigned int namescount,
       if (i > 1)
         qsort((char *)namesbuf, i, sizeof(const char *), compare_by_name);
       /* Call the callback. */
-      if (do_one(i,namesbuf,data))
+      if (do_one((unsigned int)i,namesbuf,data))
         break;
     }
   }
@@ -539,7 +539,7 @@ const char * iconv_canonicalize (const char * name)
       continue;
     }
     pool = stringpool;
-    ap = aliases_lookup(buf,bp-buf);
+    ap = aliases_lookup(buf, (unsigned int)(bp-buf));
     if (ap == NULL) {
       pool = stringpool2;
       ap = aliases2_lookup(buf);
